backend/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
gunicorn
python-dotenv
drf-yasg>=1.20.0
python-dotenv==1.0.0
whitenoise
django-modeltranslation
qrcode[pil]
